dedup and coverage
