+++ pol/api/v0/utils.py
@@ -1,0 +1,13 @@
+from typing import Dict, Optional
+
+
+def person_images(s: Optional[str]) -> Optional[Dict[str, str]]:
+    if not s:
+        return None
+
+    return {
+        "large": "https://lain.bgm.tv/pic/crt/l/" + s,
+        "medium": "https://lain.bgm.tv/pic/crt/m/" + s,
+        "small": "https://lain.bgm.tv/pic/crt/s/" + s,
+        "grid": "https://lain.bgm.tv/pic/crt/g/" + s,
+    }
